Restores old values for ballistics
I did not restore old values on energy guns, on the basis most of them did not have one to begin with.

As for balance, well, it is around as balanced as it was before the recoil rework. In a way, at least.

The placeholder multiplier based on damage was removed, so you can breathe freely now, Valo.